requirements.txt
numpy==1.24.3
pandas==2.0.3
fpdf==1.7.2
openai==0.27.8
markdown-table==2020.12.3
python-dotenv==1.0.0
PyPDF2==3.0.1